+++ packages/code-du-travail-frontend/src/modules/layout/header/HeaderSearchV2.tsx
@@ -1,0 +1,29 @@
+import { fr } from "@codegouvfr/react-dsfr";
+import Button from "@codegouvfr/react-dsfr/Button";
+import { css } from "@styled-system/css";
+
+type HeaderSearchProps = {
+  onSearchClick: () => void;
+};
+
+export const HeaderSearchV2 = ({ onSearchClick }: HeaderSearchProps) => {
+  return (
+    <div className={fr.cx("fr-header__tools", "fr-hidden", "fr-unhidden-lg")}>
+      <Button
+        id="fr-header-search-button-desktop"
+        title="Rechercher"
+        onClick={onSearchClick}
+        iconId="fr-icon-search-line"
+        iconPosition="right"
+        className={buttonStyle}
+        type="button"
+      >
+        Rechercher
+      </Button>
+    </div>
+  );
+};
+
+const buttonStyle = css({
+  height: "48px",
+});