+++ packages/code-du-travail-frontend/src/modules/recherche/SearchBar.tsx
@@ -58,6 +58,7 @@ export const SearchBar = ({ initialValue = "" }: SearchBarProps) => {
   return (
     <form role="search" onSubmit={handleSubmit} ref={formRef}>
       <Autocomplete<string>
+        id="search-bar-autocomplete"
         key={key}
         label="Recherchez sur le site, la sélection d'une option charge une nouvelle page"
         placeholder="Recherchez sur le site"
@@ -70,10 +71,10 @@ export const SearchBar = ({ initialValue = "" }: SearchBarProps) => {
             setQuery(value);
           }
         }}
-        onChange={(value, suggestions) => {
+        onChange={(value) => {
           setQuery(value ?? "");
           if (value) {
-            emitSuggestionSelectionEvent(query, value, suggestions ?? []);
+            emitSuggestionSelectionEvent(query, value);
             handleSearch(value);
           }
         }}